Move project related models to admin api slice
