+++ integration/src/test/scala/org/knora/webapi/e2e/admin/AdminFilesE2ESpec.scala
@@ -10,13 +10,13 @@ import org.apache.pekko.http.scaladsl.model.headers.BasicHttpCredentials
 import zio.ZIO
 
 import org.knora.webapi.E2ESpec
-import org.knora.webapi.messages.admin.responder.projectsmessages.PermissionCodeAndProjectRestrictedViewSettings
-import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectRestrictedViewSettingsADM
 import org.knora.webapi.messages.store.triplestoremessages.RdfDataObject
 import org.knora.webapi.messages.store.triplestoremessages.TriplestoreJsonProtocol
 import org.knora.webapi.routing.Authenticator
 import org.knora.webapi.routing.UnsafeZioRun
 import org.knora.webapi.sharedtestdata.SharedTestDataADM2
+import org.knora.webapi.slice.admin.api.model.PermissionCodeAndProjectRestrictedViewSettings
+import org.knora.webapi.slice.admin.api.model.ProjectRestrictedViewSettingsADM
 
 /**
  * End-to-End (E2E) test specification for Sipi access.
