+++ webapi/src/main/scala/org/knora/webapi/messages/util/KnoraSystemInstances.scala
@@ -6,7 +6,7 @@
 package org.knora.webapi.messages.util
 
 import org.knora.webapi.messages.admin.responder.permissionsmessages.PermissionsDataADM
-import org.knora.webapi.messages.admin.responder.projectsmessages.Project
+import org.knora.webapi.slice.admin.api.model.Project
 import org.knora.webapi.slice.admin.domain.model.Group
 import org.knora.webapi.slice.admin.domain.model.KnoraUser
 import org.knora.webapi.slice.admin.domain.model.User
