+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/KnoraResponseV2.scala
@@ -10,8 +10,8 @@ import dsp.errors.BadRequestException
 import org.knora.webapi._
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.messages.OntologyConstants
-import org.knora.webapi.messages.admin.responder.projectsmessages.Project
 import org.knora.webapi.messages.util.rdf._
+import org.knora.webapi.slice.admin.api.model.Project
 import org.knora.webapi.slice.common.api.KnoraResponseRenderer
 
 /**
