+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/FilesEndpoints.scala
@@ -11,10 +11,10 @@ import sttp.tapir.generic.auto.schemaForCaseClass
 import sttp.tapir.json.zio.jsonBody
 import zio.ZLayer
 
-import org.knora.webapi.messages.admin.responder.projectsmessages.PermissionCodeAndProjectRestrictedViewSettings
 import org.knora.webapi.slice.admin.api.AdminPathVariables.projectShortcode
 import org.knora.webapi.slice.admin.api.Codecs.TapirCodec.sparqlEncodedString
 import org.knora.webapi.slice.admin.api.FilesPathVar.filename
+import org.knora.webapi.slice.admin.api.model.PermissionCodeAndProjectRestrictedViewSettings
 import org.knora.webapi.slice.common.api.BaseEndpoints
 import org.knora.webapi.slice.common.domain.SparqlEncodedString
