+++ webapi/src/main/scala/org/knora/webapi/slice/admin/domain/service/KnoraProjectService.scala
@@ -12,7 +12,7 @@ import zio.ZIO
 import zio.ZLayer
 
 import dsp.errors.DuplicateValueException
-import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectIdentifierADM
+import org.knora.webapi.slice.admin.api.model.ProjectIdentifierADM
 import org.knora.webapi.slice.admin.api.model.ProjectsEndpointsRequestsAndResponses.ProjectCreateRequest
 import org.knora.webapi.slice.admin.api.model.ProjectsEndpointsRequestsAndResponses.ProjectUpdateRequest
 import org.knora.webapi.slice.admin.domain.model.KnoraProject
