+++ webapi/src/main/scala/org/knora/webapi/slice/admin/domain/service/KnoraUserService.scala
@@ -14,9 +14,9 @@ import zio.ZLayer
 
 import dsp.errors.DuplicateValueException
 import dsp.valueobjects.LanguageCode
-import org.knora.webapi.messages.admin.responder.projectsmessages.Project
 import org.knora.webapi.responders.IriService
 import org.knora.webapi.slice.admin.api.UsersEndpoints.Requests.UserCreateRequest
+import org.knora.webapi.slice.admin.api.model.Project
 import org.knora.webapi.slice.admin.domain.model.Email
 import org.knora.webapi.slice.admin.domain.model.FamilyName
 import org.knora.webapi.slice.admin.domain.model.GivenName
