+++ webapi/src/test/scala/org/knora/webapi/TestDataFactory.scala
@@ -9,8 +9,8 @@ import zio.Chunk
 import zio.NonEmptyChunk
 
 import dsp.valueobjects.LanguageCode
-import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectIdentifierADM._
 import org.knora.webapi.messages.store.triplestoremessages.StringLiteralV2
+import org.knora.webapi.slice.admin.api.model.ProjectIdentifierADM._
 import org.knora.webapi.slice.admin.domain.model.Email
 import org.knora.webapi.slice.admin.domain.model.FamilyName
 import org.knora.webapi.slice.admin.domain.model.GivenName
